requirements.txt

faiss-cpu==1.8.0.post1
Flask==3.0.3
Flask-Cors==5.0.0
langchain==0.3.1
langchain-huggingface==0.1.0
numpy==1.26.1
pandas==2.1.2
plotly==5.22.0
python-dotenv==1.0.1
requests==2.31.0
sentence-transformers==3.1.1
streamlit==1.37.1
langchain-community==0.3.1
